+++ .gitignore
@@ -1,2 +1,4 @@
 /target/
 /test-output/
+/lmsapi.log
+/allure-results/
