--- src/test/java/com/api/stepDefProgram/GetAllProgram.java
@@ -1,32 +1,0 @@
-package com.api.stepDefProgram;
-
-import io.cucumber.java.en.Given;
-import io.cucumber.java.en.Then;
-import io.cucumber.java.en.When;
-
-public class GetAllProgram {
-	@Given("User make a service request with a Base URL")
-	public void user_make_a_service_request_with_a_base_url() {
-	    // Write code here that turns the phrase above into concrete actions
-	    throw new io.cucumber.java.PendingException();
-	}
-
-	@When("GET http request is made by the User with an endpoint allPrograms")
-	public void get_http_request_is_made_by_the_user_with_an_endpoint_all_programs() {
-	    // Write code here that turns the phrase above into concrete actions
-	    throw new io.cucumber.java.PendingException();
-	}
-
-	@Then("User validate Status code as twohundred")
-	public void user_validate_status_code_as_twohundred() {
-	    // Write code here that turns the phrase above into concrete actions
-	    throw new io.cucumber.java.PendingException();
-	}
-
-	@Then("response body has {string}, {string}, {string}, {string}, {string}, {string}")
-	public void response_body_has(String string, String string2, String string3, String string4, String string5, String string6) {
-	    // Write code here that turns the phrase above into concrete actions
-	    throw new io.cucumber.java.PendingException();
-	}
-
-}
